# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FoodHub API",
        "description": "Multi-channel SaaS platform for restaurants - API Documentation",
        "contact": {
            "email": "admin@foodhub.uz"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "FoodHub API Server"
        }
    ],
    "paths": {
...
        "/api/v1/restaurants/{restaurant_slug}/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get products for specific restaurant",
                "operationId": "eeaf44a11a7a4c1751382b1839d9745f",
                "parameters": [
                    {
                        "name": "restaurant_slug",
                        "in": "path",
                        "description": "Restaurant slug",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "pizza-palace"
                        }
                    },
                    {
                        "name": "channel",
                        "in": "query",
                        "description": "Filter by channel",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "web",
                                "mobile",
                                "telegram",
                                "whatsapp",
                                "phone",
                                "pos"
                            ],
                            "example": "web"
                        }
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Filter by category ID",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Restaurant products retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Restaurant products retrieved successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "restaurant": {
                                                    "type": "object"
                                                },
                                                "products": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get product details",
                "operationId": "249e765b6489191fb8167474db8867f7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "channel",
                        "in": "query",
                        "description": "Channel for pricing",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "web",
                                "mobile",
                                "telegram",
                                "whatsapp",
                                "phone",
                                "pos"
                            ],
                            "example": "web"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Product retrieved successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "object"
                                                },
                                                "description": {
                                                    "type": "object"
                                                },
                                                "current_price": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 12.99
                                                },
                                                "restaurant": {
                                                    "type": "object"
                                                },
                                                "category": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            },
            "bearerAuth": {
                "type": "http",
                "description": "Enter token in format: Bearer {your-token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "passport": []
        }
    ]
}